As per spec, make jitter be a float instead of uint32_t despite we convert it to uint32_t in the stats

+++ worker/include/RTC/RTCP/ReceiverReport.hpp
@@ -102,9 +102,9 @@ namespace RTC
 			{
 				return uint32_t{ ntohl(this->header->jitter) };
 			}
-			void SetJitter(uint32_t jitter)
+			void SetJitter(float jitter)
 			{
-				this->header->jitter = uint32_t{ htonl(jitter) };
+				this->header->jitter = uint32_t{ htonl(static_cast<uint32_t>(jitter)) };
 			}
 			uint32_t GetLastSenderReport() const
 			{
